pyproject.toml
[tool.poetry]
name = "copier-autoupdate"
version = "0" # Will be updated by the poetry-version-plugin
description = "Copier template repo including GHA pipeline for automatically opening pull-requests on template changes."
authors = ["Wouter ten Brink <woutertenbrink@gmail.com>"]
license = "MIT"
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = "^3.9"
copier = "^9.2.0"

[tool.poetry.group.dev.dependencies]
black = "^24.4.2"
flake8 = "^7.0.0"
flake8-isort = "^6.1.1"
flake8-pyproject = "^1.2.3"
pep8-naming = "^0.13.3"
pytest = "^8.2.0"
PyYAML = "^6.0.1"
pre-commit = "^3.7.0"

[tool.flake8]
extend-exclude = [
    ".venv/"
]
max-line-length = 120

[tool.pytest.ini_options]
testpaths = [
    "tests"
]

[tool.poetry-git-version-plugin]
make_alpha_version = true

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
